+++ src/Service/OS2Forms/SubmissionNormalizers/AbstractSubmissionNormalizer.php
@@ -112,7 +112,9 @@ abstract class AbstractSubmissionNormalizer implements SubmissionNormalizerInter
                     // Create document
                     $document = $this->documentUploader->createDocumentFromPath($newFilePath, basename($documentUrl), 'OS2Forms');
                 } finally {
-                    unlink($newFilePath);
+                    if (file_exists($newFilePath)) {
+                        unlink($newFilePath);
+                    }
                 }
 
                 $documents[] = $document;
